backend/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
google-auth>=2.23.0
django-cors-headers>=4.3.0
gunicorn